backend/requirements.txt
Flask==3.1.*
flask-cors==5.0.*
google-genai>=1.0
python-dotenv==1.*
